+++ core/services/keystore/models.go
@@ -5,6 +5,8 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/smartcontractkit/chainlink/core/services/keystore/keys/ocr2key"
+
 	gethkeystore "github.com/ethereum/go-ethereum/accounts/keystore"
 	"github.com/pkg/errors"
 	"github.com/smartcontractkit/chainlink/core/logger"
@@ -69,20 +71,22 @@ func (ks keyStates) validate(kr keyRing) (err error) {
 }
 
 type keyRing struct {
-	CSA map[string]csakey.KeyV2
-	Eth map[string]ethkey.KeyV2
-	OCR map[string]ocrkey.KeyV2
-	P2P map[string]p2pkey.KeyV2
-	VRF map[string]vrfkey.KeyV2
+	CSA  map[string]csakey.KeyV2
+	Eth  map[string]ethkey.KeyV2
+	OCR  map[string]ocrkey.KeyV2
+	OCR2 map[string]ocr2key.KeyBundle
+	P2P  map[string]p2pkey.KeyV2
+	VRF  map[string]vrfkey.KeyV2
 }
 
 func newKeyRing() keyRing {
 	return keyRing{
-		CSA: make(map[string]csakey.KeyV2),
-		Eth: make(map[string]ethkey.KeyV2),
-		OCR: make(map[string]ocrkey.KeyV2),
-		P2P: make(map[string]p2pkey.KeyV2),
-		VRF: make(map[string]vrfkey.KeyV2),
+		CSA:  make(map[string]csakey.KeyV2),
+		Eth:  make(map[string]ethkey.KeyV2),
+		OCR:  make(map[string]ocrkey.KeyV2),
+		OCR2: make(map[string]ocr2key.KeyBundle),
+		P2P:  make(map[string]p2pkey.KeyV2),
+		VRF:  make(map[string]vrfkey.KeyV2),
 	}
 }
 
@@ -124,6 +128,9 @@ func (kr *keyRing) raw() (rawKeys rawKeyRing) {
 	}
 	for _, vrfKey := range kr.VRF {
 		rawKeys.VRF = append(rawKeys.VRF, vrfKey.Raw())
+	}
+	for _, ocr2key := range kr.OCR2 {
+		rawKeys.OCR2 = append(rawKeys.OCR2, ocr2key.Raw())
 	}
 	return rawKeys
 }
@@ -171,11 +178,12 @@ func (kr *keyRing) logPubKeys(lggr logger.Logger) {
 // it holds only the essential key information to avoid adding unecessary data
 // (like public keys) to the database
 type rawKeyRing struct {
-	Eth []ethkey.Raw
-	CSA []csakey.Raw
-	OCR []ocrkey.Raw
-	P2P []p2pkey.Raw
-	VRF []vrfkey.Raw
+	Eth  []ethkey.Raw
+	CSA  []csakey.Raw
+	OCR  []ocrkey.Raw
+	OCR2 []ocr2key.Raw
+	P2P  []p2pkey.Raw
+	VRF  []vrfkey.Raw
 }
 
 func (rawKeys rawKeyRing) keys() (keyRing, error) {
@@ -191,6 +199,10 @@ func (rawKeys rawKeyRing) keys() (keyRing, error) {
 	for _, rawOCRKey := range rawKeys.OCR {
 		ocrKey := rawOCRKey.Key()
 		keyRing.OCR[ocrKey.ID()] = ocrKey
+	}
+	for _, rawOCR2Key := range rawKeys.OCR2 {
+		ocr2Key := rawOCR2Key.Key()
+		keyRing.OCR2[ocr2Key.ID()] = ocr2Key
 	}
 	for _, rawP2PKey := range rawKeys.P2P {
 		p2pKey := rawP2PKey.Key()