+++ core/services/offchainreporting/validate.go
@@ -54,7 +54,6 @@ func ValidatedOracleSpecToml(chainSet evm.ChainSet, tomlString string) (job.Job,
 	}
 	jb.OffchainreportingOracleSpec = &spec
 
-	// TODO(#175801038): upstream support for time.Duration defaults in go-toml
 	if jb.Type != job.OffchainReporting {
 		return jb, errors.Errorf("the only supported type is currently 'offchainreporting', got %s", jb.Type)
 	}
@@ -156,7 +155,6 @@ func validateExplicitlySetKeys(tree *toml.Tree, expected map[string]struct{}, no
 	var err error
 	// top level keys only
 	for _, k := range tree.Keys() {
-		// TODO(#175801577): upstream a way to check for children in go-toml
 		if _, ok := notExpected[k]; ok {
 			err = multierr.Append(err, errors.Errorf("unrecognised key for %s peer: %s", peerType, k))
 		}