+++ operator_ui/src/pages/Dashboards/Index.tsx
@@ -1,95 +1,39 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import Grid from '@material-ui/core/Grid'
-import Activity from 'components/Dashboards/Activity'
+
 import TokenBalanceCard from 'components/Cards/TokenBalance'
 import Footer from 'components/Footer'
 import Content from 'components/Content'
 import { fetchAccountBalance } from 'actionCreators'
 import accountBalanceSelector from 'selectors/accountBalance'
-import RecentlyCreated from 'components/Jobs/RecentlyCreated'
-import { v2 } from 'api'
-import { JobRunV2, Resource } from 'core/store/models'
+
+import { Activity } from 'src/screens/Dashboard/Activity'
+import { RecentJobs } from 'src/screens/Dashboard/RecentJobs'
 
 type Props = {
   recentJobRunsCount: number
   recentlyCreatedPageSize: number
 }
 
-const fetchJobs = async (pageSize: number) => {
-  const jobs = await v2.jobs.getJobSpecs()
-
-  return jobs.data
-    .reverse()
-    .slice(0, pageSize)
-    .map(({ attributes, id }) => {
-      let createdAt = ''
-
-      if (attributes.directRequestSpec) {
-        createdAt = attributes.directRequestSpec.createdAt
-      } else if (attributes.fluxMonitorSpec) {
-        createdAt = attributes.fluxMonitorSpec.createdAt
-      } else if (attributes.offChainReportingOracleSpec) {
-        createdAt = attributes.offChainReportingOracleSpec.createdAt
-      } else if (attributes.keeperSpec) {
-        createdAt = attributes.keeperSpec.createdAt
-      } else if (attributes.cronSpec) {
-        createdAt = attributes.cronSpec.createdAt
-      } else if (attributes.webhookSpec) {
-        createdAt = attributes.webhookSpec.createdAt
-      } else if (attributes.vrfSpec) {
-        createdAt = attributes.vrfSpec.createdAt
-      }
-
-      return { id, createdAt, name: attributes.name }
-    })
-}
-
-const fetchRuns = async (
-  size: number,
-): Promise<{ runs: Resource<JobRunV2>[]; count: number }> => {
-  const response = await v2.runs.getAllJobRuns({ page: 1, size })
-
-  return {
-    runs: response.data as unknown as Resource<JobRunV2>[],
-    count: response.meta.count,
-  }
-}
-
 export const Index = ({
   recentJobRunsCount = 2,
   recentlyCreatedPageSize,
 }: Props) => {
-  const [jobs, setJobs] = useState<
-    { id: string; name: string | null; createdAt: string }[]
-  >([])
-  const [runs, setRuns] = useState<Resource<JobRunV2>[]>([])
-  const [count, setCount] = useState(0)
   const dispatch = useDispatch()
   const accountBalance = useSelector(accountBalanceSelector)
 
   React.useEffect(() => {
-    document.title = 'Dashboard'
-  }, [])
-
-  React.useEffect(() => {
     dispatch(fetchAccountBalance())
-    fetchJobs(recentlyCreatedPageSize).then((fetchedJobs) =>
-      setJobs(fetchedJobs),
-    )
-    fetchRuns(recentJobRunsCount).then((data) => {
-      setRuns(data.runs)
-      setCount(data.count)
-    })
   }, [dispatch, recentlyCreatedPageSize, recentJobRunsCount])
 
   return (
     <Content>
       <Grid container>
-        <Grid item xs={9}>
-          <Activity runs={runs} pageSize={recentJobRunsCount} count={count} />
+        <Grid item xs={8}>
+          <Activity />
         </Grid>
-        <Grid item xs={3}>
+        <Grid item xs={4}>
           <Grid container>
             <Grid item xs={12}>
               <TokenBalanceCard
@@ -104,7 +48,7 @@ export const Index = ({
               />
             </Grid>
             <Grid item xs={12}>
-              <RecentlyCreated jobs={jobs} />
+              <RecentJobs />
             </Grid>
           </Grid>
         </Grid>